Distributing blocs hierarchically Adding main page widget for folder Moving theme folder
