+++ lib/pages/help_selector/help_list/help_list.dart
@@ -1,4 +1,5 @@
 export 'bloc/bloc.dart';
-export 'widgets/widgets.dart';
-export 'model/model.dart';
 export 'chat/chat.dart';
+export 'help_list_page.dart';
+export 'model/model.dart';
+export 'widgets/widgets.dart';
