+++ lib/pages/help_selector/help_list/widgets/widgets.dart
@@ -1,3 +1,2 @@
 export 'help_list_footer.dart';
 export 'help_card.dart';
-export 'help_list_page.dart';
