+++ lib/pages/help_selector/help_selector.dart
@@ -1,4 +1,5 @@
 export 'bloc/bloc.dart';
-export 'widgets/widgets.dart';
 export 'help_list/help_list.dart';
+export 'help_selector_page.dart';
 export 'need_help/need_help.dart';
+export 'widgets/widgets.dart';
