+++ lib/pages/help_selector/need_help/need_help.dart
@@ -1,2 +1,3 @@
 export 'bloc/bloc.dart';
+export 'need_help_page.dart';
 export 'widgets/widgets.dart';
