+++ lib/pages/help_selector/widgets/widgets.dart
@@ -1,5 +1,4 @@
 export 'action_footer.dart';
-export 'help_selector_page.dart';
 export 'need_help.dart';
 export 'wanna_help.dart';
 export 'help_selector_widget.dart';
