+++ lib/pages/home/bloc/home_bloc.dart
@@ -7,10 +7,7 @@ import 'package:handyclientapp/services/device_info_service.dart';
 import 'home_event.dart';
 import 'home_state.dart';
 
-
-
 class HomeBloc extends Bloc<HomeEvent, HomeState> {
-
   final DeviceInfoService deviceInfoService;
   HomeBloc({@required this.deviceInfoService});
 
@@ -23,13 +20,13 @@ class HomeBloc extends Bloc<HomeEvent, HomeState> {
   Stream<HomeState> mapEventToState(
     HomeEvent event,
   ) async* {
-      if(event is HomeInitializeEvent){
-        final deviceInfo = await deviceInfoService.getDeviceInfo();
-        if (deviceInfo.hasLoggedIn) {
-          yield HomeIntroState();
-        } else {
-          yield HomeHelpSelectState();
-        }
+    if (event is HomeInitializeEvent) {
+      final deviceInfo = await deviceInfoService.getDeviceInfo();
+      if (!deviceInfo.hasLoggedIn) {
+        yield HomeIntroState();
+      } else {
+        yield HomeHelpSelectState();
       }
+    }
   }
 }
