+++ lib/pages/home/intro/intro.dart
@@ -1,3 +1,4 @@
 export 'bloc/bloc.dart';
+export 'intro_page.dart';
 export 'model/model.dart';
 export 'widgets/widgets.dart';
