+++ lib/pages/home/intro/widgets/widgets.dart
@@ -1,3 +1,2 @@
 export 'intro_card.dart';
-export 'intro_page.dart';
 export 'intro_card_list.dart';
